added download adn s3 storage functionality

+++ src/app.ts
@@ -9,11 +9,10 @@ import queryVersionRoutes from './apis/queryVersion.js';
 import { fetchVersion } from './tools/fetchVersion.js';
 import { searchPackages, searchPackagesRDS } from './tools/searchPackages.js';
 import { contentToURL, urlToContent } from './apis/helpers.js';
-import { storePackage, readAllPackages, readPackage, readPackageRating, deleteAllPackages } from './rds/index.js';
+import { storePackage, readAllPackages, readPackage, readPackageRating, deleteAllPackages, downloadPackageContent } from './rds/index.js';
 import cors from 'cors';
 import bodyParser from 'body-parser';
-import { uploadToS3 } from './tools/uploadToS3.js';
-import { ingestPackageHelper } from './tools/ingest.js';
+import { readFromS3, deleteFromS3 } from './tools/uploadToS3.js';
 
 const app: Application = express();
 const port = 8081;
@@ -96,6 +95,16 @@ app.post('/packages', async (req: Request, res: Response) => { //works
 });
 
 app.delete('/reset', async (req: Request, res: Response) => { //works
+  const packages = await readAllPackages();
+  if (!packages) {
+    return res.status(500).send("Failed to read packages from RDS.");
+  }
+
+  for (const pkg of packages) {
+    let path = `${pkg.metadata.Name}/${pkg.metadata.ID}`;
+    await deleteFromS3(path);
+  }
+
   const result = await deleteAllPackages();
 
   if (!result) {
@@ -122,8 +131,10 @@ app.get('/package/:id', async (req: Request, res: Response) => {
     }
 
     try {
-      const content = await urlToContent(pkg.data.URL);
-      if (content === 'Failed to get the zip file') {
+      //const content = await urlToContent(pkg.data.URL);
+      const s3path = `${pkg.metadata.Name}/${pkg.metadata.ID}`;
+      const content = await readFromS3(s3path);
+      if (content === undefined) {
         return res.status(500).send("Failed to retrieve content.");
       }
       pkg.data.Content = content;
@@ -170,6 +181,7 @@ app.post('/package/:id', async (req: Request, res: Response) => { //update this 
   }
   
   const pkg = await readPackage(req.params.id);
+  //const pkgContent = await readFromS3(req.params.id);
 
   if (!pkg) {
     return res.status(404).send("Package does not exist.");
@@ -275,20 +287,10 @@ app.post('/package', async (req: Request, res: Response) => {
       return res.status(424).send("Package is not uploaded due to the disqualified rating.");
     }
   });
-  
-  let s3result = null;
+
   const result = await storePackage(newPackage, JSON.parse(scores));
-  if (req.body.Content) {
-    s3result = await uploadToS3(ID, req.body.Content);
-  } else {
-    s3result = await ingestPackageHelper(ID, repo, version);
-  }
-  if (!result || !s3result) {
-    if (s3result) {
-      return res.status(500).send("Failed to store package data.");
-    } else {
-      return res.status(500).send("Failed to upload to S3.");
-    }
+  if (!result) {
+    return res.status(500).send("Failed to store package.");
   }
 
   return res.status(201).json(newPackage);
@@ -350,6 +352,32 @@ app.get('/package/:id/cost', async (req: Request, res: Response) => { //works
   pkgCost[pkg.metadata.ID].totalCost = parseFloat(pkgCost[pkg.metadata.ID].totalCost.toFixed(2));
 
   res.status(200).json(pkgCost);
+});
+
+app.get('/package/:id/download', async (req: Request, res: Response) => {
+  if (!req.params.id) {
+    return res.status(400).send("There is missing field(s) in the PackageID or it is formed improperly, or is invalid.");
+  }
+
+  const pkg = await readPackage(req.params.id);
+
+  if (!pkg) {
+    return res.status(404).send("Package does not exist.");
+  }
+
+  if (!pkg.data.URL) {
+    return res.status(400).send("There is missing field(s) in the PackageData or it is formed improperly, or is invalid.");
+  }
+
+  const response = await downloadPackageContent(pkg.metadata.ID);
+
+  if (response === null) {
+    return res.status(500).send("Failed to retrieve content.");
+  } else if (response.includes('Failed')) {
+      return res.status(500).send(response);
+  }
+  return res.status(200).send(response);
+
 });
 
 app.get('/tracks', (req: Request, res: Response) => {
